# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Omniviv API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/stations": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "List all stations that have platforms linked to them, optionally filtered by bounding box",
        "operationId": "list_stations",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "description": "Bounding box: west,south,east,north (comma-separated)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of stations with their platforms and stop positions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/stations/{osm_id}": {
      "get": {
        "tags": [
          "stations"
        ],
        "summary": "Get a single station by its OSM ID",
        "operationId": "get_station",
        "parameters": [
          {
            "name": "osm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Station details with platforms and stop positions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Station"
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Station": {
        "type": "object",
        "required": [
          "osm_id",
          "osm_type",
          "lat",
          "lon",
          "min_zoom",
          "platforms",
          "stop_positions",
          "platform_ways"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "min_zoom": {
            "type": "integer",
            "format": "int32",
            "description": "Minimum zoom level at which this station should be shown (6=rail, 10=tram, 13=bus)"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "osm_type": {
            "type": "string"
          },
          "platform_ways": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationPlatformWay"
            }
          },
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationPlatform"
            }
          },
          "ref_ifopt": {
            "type": [
              "string",
              "null"
            ]
          },
          "stop_positions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationStopPosition"
            }
          }
        }
      },
      "StationListResponse": {
        "type": "object",
        "required": [
          "stations"
        ],
        "properties": {
          "stations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Station"
            }
          }
        }
      },
      "StationPlatform": {
        "type": "object",
        "description": "Platform info nested in station response",
        "required": [
          "osm_id",
          "lat",
          "lon",
          "gtfs_stop_ids"
        ],
        "properties": {
          "gtfs_stop_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "GTFS stop IDs matched to this platform via spatial matching"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "ref_ifopt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "StationPlatformWay": {
        "type": "object",
        "description": "Platform way info (physical platform outline centroid) nested in station response",
        "required": [
          "osm_id",
          "lat",
          "lon"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "ref_ifopt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "StationStopPosition": {
        "type": "object",
        "description": "Stop position info nested in station response",
        "required": [
          "osm_id",
          "lat",
          "lon",
          "gtfs_stop_ids"
        ],
        "properties": {
          "gtfs_stop_ids": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "GTFS stop IDs matched to this stop position via spatial matching"
          },
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "osm_id": {
            "type": "integer",
            "format": "int64"
          },
          "platform_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "ref": {
            "type": [
              "string",
              "null"
            ]
          },
          "ref_ifopt": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  },
...
}